+++ app/controllers/alchemy/api/pages_controller.rb
@@ -12,6 +12,22 @@ module Alchemy
       respond_with @pages
     end
 
+    # Returns all pages as nested json object for tree views
+    #
+    # Pass a page_id param to only load tree for this page
+    #
+    # Pass elements=true param to include elements for pages
+    #
+    def nested
+      @page = Page.find_by(id: params[:page_id]) || Language.current_root_page
+
+      render json: PageTreeSerializer.new(@page,
+        ability: current_ability,
+        user: current_alchemy_user,
+        elements: params[:elements],
+        full: true)
+    end
+
     # Returns a json object for page
     #
     # You can either load the page via id or its urlname